+++ packages/client/verifiable-credentials/src/crossmintAPI.ts
@@ -35,15 +35,27 @@ class CrossmintAPI {
         return this.environment;
     }
 
-    public getHeaders() {
+    isProd() {
+        return this.getEnvironment() === "prod" || this.getEnvironment() === "production";
+    }
+
+    public getHeaders(post = false) {
         if (!this.apiKey) {
             throw new Error("Credentials not set");
         }
 
-        return {
+        const headers = {
             "x-api-key": this.apiKey,
             accept: "application/json",
         };
+
+        if (post) {
+            return {
+                ...headers,
+                "Content-Type": "application/json",
+            };
+        }
+        return headers;
     }
 
     public getBaseUrl() {
@@ -65,7 +77,7 @@ class CrossmintAPI {
     public init(
         apiKey: string,
         config: {
-            environment: "staging" | "prod" | string;
+            environment: "staging" | "prod" | "production" | string;
             ipfsGateways?: string[];
             ipfsTimeout?: number;
             blockchainRpcs?: ChainRPCConfig;