+++ packages/client/verifiable-credentials/src/index.ts
@@ -18,16 +18,18 @@ export type {
 export { verifyCredential } from "./verifiableCredentialsSDK";
 // Presentation
 export {
-    getNFTFromLocator,
+    getCredentialNFTFromLocator,
     ContractMetadataService,
-    // getUsersCredentialNfts, // Wrapped
+    // getCredentialNfts, // Wrapped
     // CredentialService, // Wrapped
 } from "./verifiableCredentialsSDK";
 
-export { getUsersCredentialNfts } from "./presentation/getCredentialNfts";
+export { getCredentialNfts } from "./presentation/getCredentialNfts";
 export { CredentialService } from "./presentation/getCredential";
 // Decryption
-export { Lit } from "./services/lit";
+export { Lit } from "./decryption/lit";
+export { WalletAuthService } from "./services/walletAuth";
+export { CrossmintDecrypt, CrossmintMetamaskDecrypt } from "./decryption/wallet";
 
 // Utils
 export {