+++ apps/payments/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-vue-ui-starter
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies [c5f8027]
+    -   @crossmint/client-sdk-base@1.2.0
+    -   @crossmint/client-sdk-vue-ui@1.1.10
+
 ## 1.1.9
 
 ### Patch Changes