+++ packages/client/ui/react-ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @crossmint/client-sdk-react-ui
 
+## 1.2.1
+
+### Patch Changes
+
+-   65bcbee: embed: fiat fix live updating params
+
+## 1.2.0
+
+### Minor Changes
+
+-   c5f8027: Bump @solana/web3.js
+
+### Patch Changes
+
+-   Updated dependencies [c5f8027]
+    -   @crossmint/client-sdk-base@1.2.0
+
 ## 1.1.11
 
 ### Patch Changes