+++ packages/client/ui/vanilla-ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @crossmint/client-sdk-vanilla-ui
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies [c5f8027]
+    -   @crossmint/client-sdk-base@1.2.0
+
 ## 1.1.9
 
 ### Patch Changes